Add files via upload
Given a Parking lot which can hold n Cars. Every car is issued a ticket for a slot
and the slot is assigned based on the nearest to the entry. The system also
returns some queries such as:
- Registration numbers of all cars of a particular colour.
- Slot number in which a car with a given registration number is parked.
- Slot numbers of all slots where a car of a particular colour is parked.
Write functional test cases for the parking lot system. Please include happy path
and edge case test scenarios. Would be preferred if you can mark the test cases
based on their priority. (p0 - p3)